refactor(html): tokenizer state
